requirements.txt
django-ninja==0.16.1
pytest==6.2.5
httpx==0.21.0